chore(keycloakService): MGDSTRM-2350 Add sentry capture to keycloak service

+++ pkg/services/keycloak.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/Nerzal/gocloak/v8"
+	"github.com/getsentry/sentry-go"
 	"github.com/golang/glog"
 
 	"github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pkg/api"
@@ -59,7 +60,10 @@ func NewKeycloakService(config *config.KeycloakConfig, realmConfig *config.Keycl
 }
 
 func (kc *keycloakService) RegisterKafkaClientInSSO(kafkaClusterName string, orgId string) (string, *errors.ServiceError) {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		return "", errors.GeneralError("failed to get access token for the sso client: %v", tokenErr)
+	}
 	internalClientId, err := kc.kcClient.IsClientExist(kafkaClusterName, accessToken)
 	if err != nil {
 		return "", errors.GeneralError("failed to check the sso client exists:%v", err)
@@ -95,11 +99,13 @@ func (kc *keycloakService) RegisterKafkaClientInSSO(kafkaClusterName string, org
 func (kc *keycloakService) RegisterOSDClusterClientInSSO(clusterId string, clusterOathCallbackURI string) (string, *errors.ServiceError) {
 	accessToken, tokenErr := kc.kcClient.GetToken()
 	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
 		return "", errors.GeneralError("failed to get token for the sso client: %v", tokenErr)
 	}
 
 	internalClientId, err := kc.kcClient.IsClientExist(clusterId, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return "", errors.GeneralError("failed to check the sso client exists: %v", err)
 	}
 
@@ -120,10 +126,12 @@ func (kc *keycloakService) RegisterOSDClusterClientInSSO(clusterId string, clust
 	clientConfig := kc.kcClient.ClientConfig(c)
 	internalClient, err := kc.kcClient.CreateClient(clientConfig, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return "", errors.FailedToCreateSSOClient("failed to create the sso client: %v", err)
 	}
 	secretValue, err := kc.kcClient.GetClientSecret(internalClient, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return "", errors.FailedToGetSSOClientSecret("failed to get the sso client secret: %v", err)
 	}
 
@@ -131,13 +139,18 @@ func (kc *keycloakService) RegisterOSDClusterClientInSSO(clusterId string, clust
 }
 
 func (kc *keycloakService) DeRegisterClientInSSO(clientId string) *errors.ServiceError {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return errors.GeneralError("failed to get access token to deregister OSD cluster client in sso: %v", tokenErr)
+	}
 	internalClientID, _ := kc.kcClient.IsClientExist(clientId, accessToken)
 	if internalClientID == "" {
 		return nil
 	}
 	err := kc.kcClient.DeleteClient(internalClientID, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return errors.FailedToDeleteSSOClient("failed to delete the sso client: %v", err)
 	}
 	return nil
@@ -152,7 +165,10 @@ func (kc *keycloakService) GetRealmConfig() *config.KeycloakRealmConfig {
 }
 
 func (kc keycloakService) IsKafkaClientExist(clientId string) *errors.ServiceError {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		return errors.GeneralError("failed to get access token to check kafka clients' existance: %v", tokenErr)
+	}
 	_, err := kc.kcClient.IsClientExist(clientId, accessToken)
 	if err != nil {
 		return errors.FailedToGetSSOClient("failed to get the sso client: %v", err)
@@ -175,9 +191,14 @@ func (kc keycloakService) GetKafkaClientSecret(clientId string) (string, *errors
 
 func (kc *keycloakService) CreateServiceAccount(serviceAccountRequest *api.ServiceAccountRequest, ctx context.Context) (*api.ServiceAccount, *errors.ServiceError) {
 	var serviceAcc api.ServiceAccount
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return nil, errors.GeneralError("failed to get access token: %v", tokenErr)
+	}
 	claims, err := auth.GetClaimsFromContext(ctx)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.Unauthenticated("user not authenticated")
 	}
 	orgId := auth.GetOrgIdFromClaims(claims)
@@ -213,20 +234,24 @@ func (kc *keycloakService) CreateServiceAccount(serviceAccountRequest *api.Servi
 	clientConfig := kc.kcClient.ClientConfig(c)
 	internalClient, err := kc.kcClient.CreateClient(clientConfig, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.FailedToCreateServiceAccount("failed to create the service account: %v", err)
 	}
 	clientSecret, err := kc.kcClient.GetClientSecret(internalClient, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.FailedToGetSSOClientSecret("failed to get service account secret: %v", err)
 	}
 	serviceAccountUser, err := kc.kcClient.GetClientServiceAccount(accessToken, internalClient)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.FailedToGetServiceAccount("failed fetch the service account user: %v", err)
 	}
 	serviceAccountUser.Attributes = &rhAccountID
 	serAccUser := *serviceAccountUser
 	updateErr := kc.kcClient.UpdateServiceAccountUser(accessToken, serAccUser)
 	if updateErr != nil {
+		sentry.CaptureException(updateErr)
 		return nil, errors.GeneralError("failed add attributes to service account user: %v", updateErr)
 	}
 	serviceAcc.ID = internalClient
@@ -247,15 +272,21 @@ func (kc *keycloakService) buildServiceAccountIdentifier() string {
 }
 
 func (kc *keycloakService) ListServiceAcc(ctx context.Context, first int, max int) ([]api.ServiceAccount, *errors.ServiceError) {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return nil, errors.GeneralError("failed to get access token: %v", tokenErr)
+	}
 	claims, err := auth.GetClaimsFromContext(ctx)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.Unauthenticated("user not authenticated")
 	}
 	orgId := auth.GetOrgIdFromClaims(claims)
 	var sa []api.ServiceAccount
 	clients, err := kc.kcClient.GetClients(accessToken, first, max)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.GeneralError("failed to check the sso client exists: %v", err)
 	}
 	for _, client := range clients {
@@ -280,32 +311,45 @@ func (kc *keycloakService) ListServiceAcc(ctx context.Context, first int, max in
 }
 
 func (kc *keycloakService) DeleteServiceAccount(ctx context.Context, id string) *errors.ServiceError {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return errors.GeneralError("failed to get access token to delete service account: %v", tokenErr)
+	}
 	claims, err := auth.GetClaimsFromContext(ctx)
 	if err != nil {
+		sentry.CaptureException(err)
 		return errors.Unauthenticated("user not authenticated")
 	}
 	orgId := auth.GetOrgIdFromClaims(claims)
 	userId := auth.GetAccountIdFromClaims(claims)
 	c, err := kc.kcClient.GetClientById(id, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return errors.FailedToGetServiceAccount("failed to check the service account exists: %v", err)
 	}
 	if kc.kcClient.IsSameOrg(c, orgId) && kc.kcClient.IsOwner(c, userId) {
 		err = kc.kcClient.DeleteClient(id, accessToken)
 		if err != nil {
+			sentry.CaptureException(err)
 			return errors.FailedToDeleteServiceAccount("failed to delete service account: %v", err)
 		}
 		return nil
 	} else {
+		sentry.CaptureException(err)
 		return errors.Forbidden("can not delete sso client due to permission error")
 	}
 }
 
 func (kc *keycloakService) ResetServiceAccountCredentials(ctx context.Context, id string) (*api.ServiceAccount, *errors.ServiceError) {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return nil, errors.GeneralError("failed to get access token to reset service account credentials: %v", tokenErr)
+	}
 	claims, err := auth.GetClaimsFromContext(ctx)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.Unauthenticated("user not authenticated")
 	}
 	orgId := auth.GetOrgIdFromClaims(claims)
@@ -313,11 +357,13 @@ func (kc *keycloakService) ResetServiceAccountCredentials(ctx context.Context, i
 	owner := auth.GetUsernameFromClaims(claims)
 	c, err := kc.kcClient.GetClientById(id, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.FailedToGetServiceAccount("failed to check the service account exists: %v", err)
 	}
 	if kc.kcClient.IsSameOrg(c, orgId) && kc.kcClient.IsOwner(c, userId) {
 		credRep, err := kc.kcClient.RegenerateClientSecret(accessToken, id)
 		if err != nil {
+			sentry.CaptureException(err)
 			return nil, errors.GeneralError("failed to regenerate service account secret: %v", err)
 		}
 		value := *credRep.Value
@@ -342,9 +388,14 @@ func (kc *keycloakService) ResetServiceAccountCredentials(ctx context.Context, i
 }
 
 func (kc *keycloakService) GetServiceAccountById(ctx context.Context, id string) (*api.ServiceAccount, *errors.ServiceError) {
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return nil, errors.GeneralError("failed to get access token to retrieve service account by id: %v", tokenErr)
+	}
 	claims, err := auth.GetClaimsFromContext(ctx)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.Unauthenticated("user not authenticated")
 	}
 	orgId := auth.GetOrgIdFromClaims(claims)
@@ -352,6 +403,7 @@ func (kc *keycloakService) GetServiceAccountById(ctx context.Context, id string)
 	owner := auth.GetUsernameFromClaims(claims)
 	c, err := kc.kcClient.GetClientById(id, accessToken)
 	if err != nil {
+		sentry.CaptureException(err)
 		return nil, errors.FailedToGetServiceAccount("failed to check the service account exists: %v", err)
 	}
 	attributes := c.Attributes
@@ -398,8 +450,11 @@ func (kc *keycloakService) RegisterConnectorFleetshardOperatorServiceAccount(age
 }
 
 func (kc *keycloakService) registerAgentServiceAccount(clusterId string, serviceAccountId string, agentClusterId string, roleName string) (*api.ServiceAccount, *errors.ServiceError) {
-
-	accessToken, _ := kc.kcClient.GetToken()
+	accessToken, tokenErr := kc.kcClient.GetToken()
+	if tokenErr != nil {
+		sentry.CaptureException(tokenErr)
+		return nil, errors.GeneralError("failed to get access token to register agent service account: %v", tokenErr)
+	}
 	role, err := kc.createRealmRoleIfNotExists(accessToken, roleName)
 	if err != nil {
 		return nil, err
